InternetandCloud/GCP_API_APP/requirements.txt
Flask==1.1.1
gunicorn==19.9.0
google-cloud-vision==0.39.0
google-cloud-storage==1.19.1
google-cloud-datastore==1.9.0
google-cloud-translate==1.3.1
google-cloud-texttospeech==0.5.0
google-cloud-videointelligence==1.11.0
google-cloud-speech==1.2.0
google-cloud-language==1.3.0

